patch: extract all settings into one module
