+++ src/noise_helpers_64.rs
@@ -1,6 +1,7 @@
 use simdeez::Simd;
 
 use super::NoiseType;
+use crate::dimensional_being::DimensionalBeing;
 
 use crate::noise::cell2_64::{cellular2_2d, cellular2_3d};
 use crate::noise::cell_64::{cellular_2d, cellular_3d};
@@ -14,7 +15,7 @@ use std::f64;
 macro_rules! get_1d_noise_helper_f64  {
     ($Setting:expr,$f:expr $(,$arg:expr)*) => {
  {
-    let dim = $Setting.dim;
+    let dim = $Setting.get_dimensions();
     let freq_x = S::set1_pd($Setting.freq_x as f64);
     let start_x = dim.x as f64;
     let width = dim.width;
@@ -73,7 +74,7 @@ macro_rules! get_1d_noise_helper_f64  {
 
 macro_rules! get_2d_noise_helper_f64 {
     ($Setting:expr,$f:expr $(,$arg:expr)*)=> {{
-    let dim = $Setting.dim;
+    let dim = $Setting.get_dimensions();
     let freq_x = S::set1_pd($Setting.freq_x as f64);
     let freq_y = S::set1_pd($Setting.freq_y as f64);
     let start_x = dim.x as f64;
@@ -138,7 +139,7 @@ macro_rules! get_2d_noise_helper_f64 {
 
 macro_rules! get_3d_noise_helper_f64 {
     ($Setting:expr,$f:expr $(,$arg:expr)*) => {{
-    let dim = $Setting.dim;
+    let dim = $Setting.get_dimensions();
     let freq_x = S::set1_pd($Setting.freq_x as f64);
     let freq_y = S::set1_pd($Setting.freq_y as f64);
     let freq_z = S::set1_pd($Setting.freq_z as f64);
@@ -210,7 +211,7 @@ macro_rules! get_3d_noise_helper_f64 {
 
 macro_rules! get_4d_noise_helper_f64 {
     ($Setting:expr,$f:expr $(,$arg:expr)*) => {{
-    let dim = $Setting.dim;
+    let dim = $Setting.get_dimensions();
     let freq_x = S::set1_pd($Setting.freq_x as f64);
     let freq_y = S::set1_pd($Setting.freq_y as f64);
     let freq_z = S::set1_pd($Setting.freq_z as f64);
@@ -297,7 +298,7 @@ pub unsafe fn get_1d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Ridge(s) => get_1d_noise_helper_f64!(
             s,
@@ -305,7 +306,7 @@ pub unsafe fn get_1d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Turbulence(s) => get_1d_noise_helper_f64!(
             s,
@@ -313,9 +314,11 @@ pub unsafe fn get_1d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
-        NoiseType::Gradient(s) => get_1d_noise_helper_f64!(s, simplex_1d::<S>, s.dim.seed as i64),
+        NoiseType::Gradient(s) => {
+            get_1d_noise_helper_f64!(s, simplex_1d::<S>, s.get_dimensions().seed as i64)
+        }
         NoiseType::Cellular(_) => {
             panic!("not implemented");
         }
@@ -340,7 +343,7 @@ pub unsafe fn get_2d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Ridge(s) => get_2d_noise_helper_f64!(
             s,
@@ -348,7 +351,7 @@ pub unsafe fn get_2d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Turbulence(s) => get_2d_noise_helper_f64!(
             s,
@@ -356,16 +359,18 @@ pub unsafe fn get_2d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
-        NoiseType::Gradient(s) => get_2d_noise_helper_f64!(s, simplex_2d::<S>, s.dim.seed as i64),
+        NoiseType::Gradient(s) => {
+            get_2d_noise_helper_f64!(s, simplex_2d::<S>, s.get_dimensions().seed as i64)
+        }
         NoiseType::Cellular(s) => get_2d_noise_helper_f64!(
             s,
             cellular_2d::<S>,
             s.distance_function,
             s.return_type,
             S::set1_pd(s.jitter as f64),
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Cellular2(s) => get_2d_noise_helper_f64!(
             s,
@@ -375,7 +380,7 @@ pub unsafe fn get_2d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.jitter as f64),
             s.index0,
             s.index1,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
     }
 }
@@ -395,7 +400,7 @@ pub unsafe fn get_3d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Ridge(s) => get_3d_noise_helper_f64!(
             s,
@@ -403,7 +408,7 @@ pub unsafe fn get_3d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Turbulence(s) => get_3d_noise_helper_f64!(
             s,
@@ -411,16 +416,18 @@ pub unsafe fn get_3d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
-        NoiseType::Gradient(s) => get_3d_noise_helper_f64!(s, simplex_3d::<S>, s.dim.seed as i64),
+        NoiseType::Gradient(s) => {
+            get_3d_noise_helper_f64!(s, simplex_3d::<S>, s.get_dimensions().seed as i64)
+        }
         NoiseType::Cellular(s) => get_3d_noise_helper_f64!(
             s,
             cellular_3d::<S>,
             s.distance_function,
             s.return_type,
             S::set1_pd(s.jitter as f64),
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Cellular2(s) => get_3d_noise_helper_f64!(
             s,
@@ -430,7 +437,7 @@ pub unsafe fn get_3d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.jitter as f64),
             s.index0,
             s.index1,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
     }
 }
@@ -445,7 +452,7 @@ pub unsafe fn get_4d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Ridge(s) => get_4d_noise_helper_f64!(
             s,
@@ -453,7 +460,7 @@ pub unsafe fn get_4d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
         NoiseType::Turbulence(s) => get_4d_noise_helper_f64!(
             s,
@@ -461,9 +468,11 @@ pub unsafe fn get_4d_noise_f64<S: Simd>(noise_type: &NoiseType) -> (Vec<f64>, f6
             S::set1_pd(s.lacunarity as f64),
             S::set1_pd(s.gain as f64),
             s.octaves,
-            s.dim.seed as i64
+            s.get_dimensions().seed as i64
         ),
-        NoiseType::Gradient(s) => get_4d_noise_helper_f64!(s, simplex_4d::<S>, s.dim.seed as i64),
+        NoiseType::Gradient(s) => {
+            get_4d_noise_helper_f64!(s, simplex_4d::<S>, s.get_dimensions().seed as i64)
+        }
         NoiseType::Cellular(_) => {
             panic!("not implemented");
         }
